tried to fix total number of seats

+++ database.h
@@ -18,7 +18,7 @@ public:
     QSqlQueryModel* getListOfOpenRoofStadiums();
     QSqlQueryModel* getListOfAllTeams();
     QSqlQueryModel* getSpecificTeamInfo(QString teamName);
-    long getTotalNumberOfSeats();
+    int getTotalNumberOfSeats();
     std::vector<QString> getTeamNames();
     QSqlQueryModel* getListOfStarPLayers();
     void addStadiumDistancesFromFile();
